rapd/requirements.txt
torch>=2.0
torchvision
faiss-cpu
numpy
